+++ transport.go
@@ -1,5 +1,7 @@
 package tscaddy
 
+// transport.go contains the TailscaleCaddyTransport module.
+
 import (
 	"fmt"
 	"net/http"
@@ -9,9 +11,19 @@ import (
 	"go.uber.org/zap"
 )
 
+// TailscaleCaddyTransport is a caddy transport that uses a tailscale node to make requests.
 type TailscaleCaddyTransport struct {
 	logger *zap.Logger
-	server *tsnetServerDestructor
+	node   *tailscaleNode
+}
+
+func (t *TailscaleCaddyTransport) CaddyModule() caddy.ModuleInfo {
+	return caddy.ModuleInfo{
+		ID: "http.reverse_proxy.transport.tailscale",
+		New: func() caddy.Module {
+			return new(TailscaleCaddyTransport)
+		},
+	}
 }
 
 func (t *TailscaleCaddyTransport) UnmarshalCaddyfile(d *caddyfile.Dispenser) error {
@@ -21,7 +33,8 @@ func (t *TailscaleCaddyTransport) UnmarshalCaddyfile(d *caddyfile.Dispenser) err
 func (t *TailscaleCaddyTransport) Provision(ctx caddy.Context) error {
 	t.logger = ctx.Logger()
 
-	s, err := getServer(ctx, "caddy-tsnet-client:80")
+	// TODO(will): allow users to specify a node name used to lookup that node's config in TSApp.
+	s, err := getNode(ctx, "caddy-tsnet-client")
 	if err != nil {
 		return err
 	}
@@ -34,25 +47,16 @@ func (t *TailscaleCaddyTransport) Provision(ctx caddy.Context) error {
 	if err := s.Start(); err != nil {
 		return err
 	}
-	t.server = s
+	t.node = s
 
 	return nil
-}
-
-func (t *TailscaleCaddyTransport) CaddyModule() caddy.ModuleInfo {
-	return caddy.ModuleInfo{
-		ID: "http.reverse_proxy.transport.tailscale",
-		New: func() caddy.Module {
-			return new(TailscaleCaddyTransport)
-		},
-	}
 }
 
 func (t *TailscaleCaddyTransport) RoundTrip(request *http.Request) (*http.Response, error) {
 	if request.URL.Scheme == "" {
 		request.URL.Scheme = "http"
 	}
-	return t.server.HTTPClient().Transport.RoundTrip(request)
+	return t.node.HTTPClient().Transport.RoundTrip(request)
 }
 
 var (
